endpoints de categorias y productos y su logica

+++ user-backend/user-backend/src/main/java/com/garmyshop/user_backend/repository/ColorRepository.java
@@ -3,6 +3,6 @@ package com.garmyshop.user_backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.garmyshop.user_backend.entity.Color;
 
-public interface ColorRepository extends JpaRepository<Color, Long> {
+public interface ColorRepository extends JpaRepository<Color, Integer> {
 
 }
