+++ user-backend/user-backend/src/main/java/com/garmyshop/user_backend/repository/CombinacionProductoRepository.java
@@ -3,6 +3,6 @@ package com.garmyshop.user_backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.garmyshop.user_backend.entity.CombinacionProducto;
 
-public interface CombinacionProductoRepository extends JpaRepository<CombinacionProducto, Long> {
+public interface CombinacionProductoRepository extends JpaRepository<CombinacionProducto, Integer> {
 
 }
