+++ user-backend/user-backend/src/main/java/com/garmyshop/user_backend/repository/ImagenProductoRepository.java
@@ -3,6 +3,6 @@ package com.garmyshop.user_backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.garmyshop.user_backend.entity.ImagenProducto;
 
-public interface ImagenProductoRepository extends JpaRepository<ImagenProducto, Long> {
+public interface ImagenProductoRepository extends JpaRepository<ImagenProducto, Integer> {
 
 }
