+++ user-backend/user-backend/src/main/java/com/garmyshop/user_backend/repository/TallaRepository.java
@@ -3,6 +3,6 @@ package com.garmyshop.user_backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.garmyshop.user_backend.entity.Talla;
 
-public interface TallaRepository extends JpaRepository<Talla, Long> {
+public interface TallaRepository extends JpaRepository<Talla, Integer> {
 
 }
